post modify

+++ src/app/controllers/postController.js
@@ -41,8 +41,8 @@ export const handleAddPost = async (req, res, next) => {
     const newPost = {
       post_id: generateCode,
       post_image: {
-        id: uploadedPostImage?.public_id ? uploadedPostImage?.public_id : "",
-        url: uploadedPostImage?.url ? uploadedPostImage?.url : "",
+        id: uploadedPostImage?.public_id || null,
+        url: uploadedPostImage?.url || null,
       },
       post_description: processedPost,
       post_additional: {
@@ -153,6 +153,14 @@ export const handleGetSinglePost = async (req, res, next) => {
     if (!result) {
       throw createError(404, "Post not found");
     }
+
+    const postOwner = await usersCollection.findOne(
+      {
+        user_id: result?.createdBy,
+      },
+      { projection: { name: 1, avatar: 1, _id: 0 } }
+    );
+
     // add view value
     await postsCollection.findOneAndUpdate(
       { post_id: result.post_id },
@@ -162,7 +170,7 @@ export const handleGetSinglePost = async (req, res, next) => {
     res.status(200).send({
       success: true,
       message: "Post retrieved successfully",
-      data: result,
+      data: { user_info: postOwner, ...result },
     });
   } catch (error) {
     next(error);
@@ -191,12 +199,14 @@ export const handleDeletePost = async (req, res, next) => {
       throw createError(404, "Post not found");
     }
 
-    const removeImage = await deleteFromCloudinary(
-      existingPost?.post_image?.id
-    );
+    if (existingPost?.post_image?.id) {
+      const removeImage = await deleteFromCloudinary(
+        existingPost?.post_image?.id
+      );
 
-    if (removeImage?.result != "ok") {
-      throw createError(500, "Something went wrong try again");
+      if (removeImage?.result != "ok") {
+        throw createError(500, "Something went wrong try again");
+      }
     }
 
     const result = await postsCollection.findOneAndDelete({ post_id: postId });
